# docs/architecture/openapi.yaml
openapi: 3.0.3
info:
  title: kleekTo API
  version: 1.0.0-draft
  description: |
    API продукта kleekTo — CRM для агентств недвижимости Грузии.

    Проектная версия фазы 1. Кода ещё нет; этот файл — контракт, из которого код выводится.

    ЖЁСТКИЕ ПРАВИЛА, ОТРАЖЁННЫЕ В СХЕМАХ:

    * Ни один запрос не принимает companyId, teamId или идентификатор владельца данных.
      Они берутся из сессии (правило 5). Появление такого параметра в схеме — дефект.
    * Ресурс другой компании возвращает 404, а не 403: 403 подтвердил бы существование объекта.
    * ListingPublishDraft не содержит блока owner. Не опционально — его нет в типе (правило 13).
    * Схема события об ошибке парсинга не позволяет передать персональные данные (правило 10).

    Подробности и обоснования: docs/architecture/api-contracts.md
servers:
  - url: /api/v1
    description: Версия в пути — инвариант 8
...
security:
  - bearerAuth: []
paths:
...
  /auth/refresh:
    post:
      tags: [auth]
      summary: Обновление пары токенов
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken]
              properties:
                refreshToken: { type: string }
      responses:
        '200':
          description: Новая пара, старый refresh отозван (ротация)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TokenPair' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Расширение передаёт access-токен заголовком.
        Веб использует httpOnly cookie с тем же токеном.
  responses:
...
    Unauthenticated:
      description: Токена нет, истёк или отозван
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              enum:
                - VALIDATION_ERROR
                - UNAUTHENTICATED
                - FORBIDDEN
                - NOT_FOUND
                - CONFLICT
                - DUPLICATE_BLOCKED
                - PHONE_NOT_REVEALED
                - RATE_LIMITED
                - INTERNAL
            message: { type: string }
            requestId: { type: string }
            details: { type: object }
...
    TokenPair:
      type: object
      required: [accessToken, refreshToken, expiresIn]
      properties:
        accessToken: { type: string }
        refreshToken: { type: string }
        expiresIn: { type: integer, description: Секунд до истечения access }
...